+++ src/server/puppet-implementation.ts
@@ -43,7 +43,7 @@ import {
 import { log } from '../config.js'
 import { grpcError }          from './grpc-error.js'
 import { EventStreamManager } from './event-stream-manager.js'
-import { channelPayloadToPb, channelPbToPayload, urlLinkPayloadToPb, urlLinkPbToPayload } from '../utils/pb-payload-helper.js'
+import { channelPayloadToPb, channelPbToPayload, postPbToPayload, urlLinkPayloadToPb, urlLinkPbToPayload } from '../utils/pb-payload-helper.js'
 
 function puppetImplementation (
   puppet      : PUPPET.impls.PuppetInterface,
@@ -1100,6 +1100,54 @@ function puppetImplementation (
       }
     },
 
+    getMessageBroadcastTarget: async (call, callback) => {
+      log.verbose('PuppetServiceImpl', 'getMessageBroadcastTarget()')
+
+      void call
+      try {
+        const payload = await puppet.getMessageBroadcastTarget()
+
+        const response = new grpcPuppet.GetMessageBroadcastTargetResponse()
+        response.setContactIdsList(payload.contactIds || [])
+        response.setRoomIdsList(payload.roomIds || [])
+
+        return callback(null, response)
+      } catch (e) {
+        return grpcError('getMessageBroadcastTarget', e, callback)
+      }
+    },
+
+    createMessageBroadcast: async (call, callback) => {
+      log.verbose('PuppetServiceImpl', 'createMessageBroadcast()')
+
+      try {
+        const targets = call.request.getTargetIdsList()
+        const type = call.request.getType()
+        const post = call.request.getContent()
+
+        if (!post) {
+          throw new Error('no post found')
+        }
+        if (post.getType() !== grpcPuppet.PostType.POST_TYPE_BROADCAST) {
+          throw new Error('cannot create broadcast with non-broadcast post')
+        }
+
+        const payload = postPbToPayload(post, FileBoxUuid)
+
+        const id = await puppet.createMessageBroadcast(targets, type, payload)
+
+        const response = new grpcPuppet.CreateMessageBroadcastResponse()
+        if (id) {
+          response.setId(id)
+        }
+
+        return callback(null, response)
+
+      } catch (e) {
+        return grpcError('createMessageBroadcast', e, callback)
+      }
+    },
+
     roomAdd: async (call, callback) => {
       log.verbose('PuppetServiceImpl', 'roomAdd()')
 
@@ -1764,67 +1812,7 @@ function puppetImplementation (
         if (type !== grpcPuppet.PostType.POST_TYPE_MOMENT) {
           throw new Error('cannot publish non-moment post')
         }
-        const payload: PUPPET.payloads.PostClient = {
-          type: PUPPET.types.Post.Moment,
-          sayableList: [],
-          rootId: post.getRootId(),
-          parentId: post.getParentId(),
-          visibleList: post.getVisibleListList(),
-        }
-        const sayableList = post.getSayableListList()
-        for (const sayable of sayableList) {
-          let sayablePayload: PUPPET.payloads.Sayable | undefined
-          switch (sayable.getType()) {
-            case grpcPuppet.SayableType.SAYABLE_TYPE_TEXT:
-              sayablePayload = PUPPET.payloads.sayable.text(sayable.getText() || '', sayable.getMentionIdListList())
-              break
-            case grpcPuppet.SayableType.SAYABLE_TYPE_FILE: {
-              const fileJsonStr = sayable.getFileBox()
-              if (!fileJsonStr) {
-                break
-              }
-              const file = FileBoxUuid.fromJSON(fileJsonStr)
-              sayablePayload = PUPPET.payloads.sayable.attachment(file)
-              break
-            }
-            case grpcPuppet.SayableType.SAYABLE_TYPE_URL: {
-              const urlLinkPayloadPb = sayable.getUrlLink()
-              if (!urlLinkPayloadPb) {
-                break
-              }
-              const urlLinkPayload = urlLinkPbToPayload(urlLinkPayloadPb)
-              sayablePayload = PUPPET.payloads.sayable.url(urlLinkPayload)
-              break
-            }
-            case grpcPuppet.SayableType.SAYABLE_TYPE_CHANNEL: {
-              const channelPayloadPb = sayable.getChannel()
-              if (!channelPayloadPb) {
-                break
-              }
-              const channelPayload = channelPbToPayload(channelPayloadPb!)
-              sayablePayload = PUPPET.payloads.sayable.channel(channelPayload)
-              break
-            }
-            default:
-              throw new Error(`unsupported postSayableType type ${sayable.getType()}`)
-          }
-          if (sayablePayload) {
-            payload.sayableList.push(sayablePayload)
-          } else {
-            throw new Error(`unable to fetch sayable from ${JSON.stringify(sayable.toObject())}`)
-          }
-        }
-
-        const location = post.getLocation()
-        if (location) {
-          payload.location = {
-            name: location.getName() || '',
-            accuracy: location.getAccuracy() || 15,
-            address: location.getAddress() || '',
-            latitude: location.getLatitude(),
-            longitude: location.getLatitude(),
-          }
-        }
+        const payload = postPbToPayload(post, FileBoxUuid)
 
         const momentId = await puppet.postPublish(payload)
         const response = new grpcPuppet.MomentPublishResponse()
